phase2/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
google-genai==1.2.0
httpx==0.28.0
